Allow specifying the full path of the page
Here is my use-case:

I need the host to be `0.0.0.0` so I can access the server from other
computers on the network (eg to test my website on mobile devices).
However currently if I set `host` to `0.0.0.0` my browser will open and broadcast
the request to the whole network. I might get a response from another
server. This feature would allow me to open `http://127.0.0.1` instead of `http://0.0.0.0` and avoid this issue.